COMMERCE-919 Move the Nested to the annotation package
