--- commerce-openapi-core-api/src/main/resources/com/liferay/commerce/openapi/core/nested/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
